Refactor database update logic to MainContentController and streamline silent updates on startup
